Added simplification caching to `PolygonLayer` (consistent with `PolylineLayer`) Moved `isClockwise` function into `Polygon` as a static method Added `Polygon` equality and hashing methods Added `Polygon.copyWithNewPoints` method Minor performance improvements to `Polyline` equality and hashing methods Minor syntactic changes to `PolylineLayer.build` method Removed unnecessary `CustomPainter` methods/parts from `PolylineLayer` & `PolygonLayer` Reorganized internal file structure of Polygon Layer Improved Polygon example with 138k points from GeoJson file Seperated `SimplificationToleranceSlider` in example into independent file
