+++ test/full_coverage_test.dart
@@ -17,7 +17,6 @@ import 'package:flutter_map/src/layer/general/mobile_layer_transformer.dart';
 import 'package:flutter_map/src/layer/general/translucent_pointer.dart';
 import 'package:flutter_map/src/layer/marker_layer.dart';
 import 'package:flutter_map/src/layer/overlay_image_layer.dart';
-import 'package:flutter_map/src/layer/polygon_layer/label.dart';
 import 'package:flutter_map/src/layer/polygon_layer/polygon_layer.dart';
 import 'package:flutter_map/src/layer/polyline_layer/polyline_layer.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile.dart';
